requirements.txt
uvicorn
fastapi
graphene
graphql-core<3
httpx